backend/requirements.txt
# ============================================
# Backend Requirements for Agri ROBO Tomato
# ============================================
# This file contains all Python dependencies needed for the backend API server
# Install with: pip install -r requirements.txt

# ============================================
# Web Framework & Server
# ============================================
fastapi>=0.104.1
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6

# ============================================
# Machine Learning & Deep Learning
# ============================================
# TensorFlow for model loading and inference
tensorflow>=2.10.0,<3.0.0

# Numerical computing
# Note: For Python 3.13, NumPy 2.x is required (has pre-built wheels)
# If installation fails, try: pip install numpy --only-binary :all:
numpy>=1.24.0
scipy>=1.9.0

# ============================================
# Image Processing
# ============================================
Pillow>=9.0.0

# ============================================
# Optional: Model Visualization Tools
# ============================================
# Uncomment if you need to visualize model architecture
# pydot>=1.4.2
# graphviz>=0.20.0
# Note: graphviz also requires system installation

# ============================================
# Development & Testing Tools
# ============================================
# For testing API endpoints (test_upload.py)
requests>=2.31.0

# ============================================
# Notes:
# ============================================
# - TensorFlow 2.10+ is required for Keras 3.x compatibility
# - For Raspberry Pi, consider using tensorflow-lite for better performance
# - If training models, ensure you have sufficient RAM (8GB+ recommended)
# - For production, pin exact versions for reproducibility
